REGRESSION(303851@main): [ macOS iOS Debug ] ipc/serialized-type-info.html is a constant failure https://bugs.webkit.org/show_bug.cgi?id=303876 rdar://166170583
Reviewed by Alex Christensen.

The problem is that size_t cannot be serialized. This originated
in this patch: https://commits.webkit.org/303851@main, where I
added a Vector<size_t> as a member of a struct.

I have changed this to be a Vector<uint64_t>. The type size_t cannot
be serialized for IPC because sizeof(size_t) is dependant on architecture
and watchOS sometimes communicates via IPC from a 32 bit machine to a
64 bit machine.

* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/page/Frame.cpp:
(WebCore::Frame::indexInFrameTreeSiblings const):
(WebCore::Frame::pathToFrame const):
* Source/WebCore/page/Frame.h:
* Source/WebCore/page/Page.cpp:
(WebCore::Page::findFrameByPath const):
* Source/WebCore/page/Page.h:
* Source/WebKit/Shared/WebFoundTextRange.h:
(WTF::HashTraits<WebKit::WebFoundTextRange>::constructDeletedValue):
* Source/WebKit/Shared/WebFoundTextRange.serialization.in:
* Source/WebKit/UIProcess/WebFrameProxy.cpp:
(WebKit::WebFrameProxy::childFrame const):
* Source/WebKit/UIProcess/WebFrameProxy.h:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(+[WKFoundTextRange foundTextRangeWithWebFoundTextRange:]):
(-[WKFoundDOMTextRange webFoundTextRange]):
(-[WKFoundPDFTextRange webFoundTextRange]):

Canonical link: https://commits.webkit.org/304617@main

+++ Source/WebCore/page/Frame.cpp
@@ -210,12 +210,12 @@ Ref<NavigationScheduler> Frame::protectedNavigationScheduler() const
     return m_navigationScheduler.get();
 }
 
-std::optional<size_t> Frame::indexInFrameTreeSiblings() const
+std::optional<uint64_t> Frame::indexInFrameTreeSiblings() const
 {
     if (!tree().parent())
         return std::nullopt;
 
-    for (size_t i = 0; i < tree().parent()->tree().childCount(); i++) {
+    for (uint64_t i = 0; i < tree().parent()->tree().childCount(); i++) {
         if (auto child = tree().parent()->tree().child(i); child->frameID() == this->frameID())
             return i;
     }
@@ -224,9 +224,9 @@ std::optional<size_t> Frame::indexInFrameTreeSiblings() const
     return std::nullopt;
 }
 
-Vector<size_t> Frame::pathToFrame() const
+Vector<uint64_t> Frame::pathToFrame() const
 {
-    Vector<size_t> path;
+    Vector<uint64_t> path;
     RefPtr current = this;
 
     while (current) {
